+++ internal/delivery/rest/route_boards_treasures.go
